Apply retrofit builder in services
