+++ sdk/src/main/java/com/ingresse/sdk/ClientBuilder.kt
@@ -1,13 +1,8 @@
 package com.ingresse.sdk
 
-import com.ingresse.sdk.url.builder.Host
-import com.ingresse.sdk.url.builder.URLBuilder
 import okhttp3.Interceptor
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import retrofit2.converter.scalars.ScalarsConverterFactory
 import java.util.concurrent.TimeUnit
 
 class ClientBuilder(private val client: IngresseClient, private val hasTimeout: Boolean = false) {
@@ -41,20 +36,5 @@ class ClientBuilder(private val client: IngresseClient, private val hasTimeout: 
 
             return@Interceptor chain.proceed(request)
         }
-    }
-}
-
-class RetrofitBuilder(host: Host = Host.API, private val client: IngresseClient, private val hasGsonConverter: Boolean = false, private val hasTimeout: Boolean = false) {
-    private val url = URLBuilder(host, client.environment).build()
-    private val retrofitBuilder = Retrofit.Builder()
-
-    fun build(): Retrofit { return retrofitBuilder.addComponents().build() }
-
-    private fun Retrofit.Builder.addComponents(): Retrofit.Builder {
-        if (hasGsonConverter) addConverterFactory(GsonConverterFactory.create())
-        addConverterFactory(ScalarsConverterFactory.create())
-        client(ClientBuilder(client, hasTimeout).build())
-        baseUrl(url)
-        return this
     }
 }
