+++ sdk/src/main/java/com/ingresse/sdk/services/AuthService.kt
@@ -2,6 +2,7 @@ package com.ingresse.sdk.services
 
 import com.google.gson.reflect.TypeToken
 import com.ingresse.sdk.IngresseClient
+import com.ingresse.sdk.RetrofitBuilder
 import com.ingresse.sdk.base.IngresseCallback
 import com.ingresse.sdk.base.Response
 import com.ingresse.sdk.base.RetrofitCallback
@@ -17,18 +18,13 @@ import retrofit2.Retrofit
 import retrofit2.converter.scalars.ScalarsConverterFactory
 
 class AuthService(private val client: IngresseClient) {
-    private var host = Host.API
     private var service: Auth
 
     private var mCompanyLoginCall: Call<String>? = null
     private var mRenewAuthTokenCall: Call<String>? = null
 
     init {
-        val adapter = Retrofit.Builder()
-            .addConverterFactory(ScalarsConverterFactory.create())
-            .baseUrl(URLBuilder(host, client.environment).build())
-            .build()
-
+        val adapter = RetrofitBuilder(client = client, withHttpClient = false).build()
         service = adapter.create(Auth::class.java)
     }
