+++ sdk/src/main/java/com/ingresse/sdk/services/UserService.kt
@@ -2,6 +2,7 @@ package com.ingresse.sdk.services
 
 import com.google.gson.reflect.TypeToken
 import com.ingresse.sdk.IngresseClient
+import com.ingresse.sdk.RetrofitBuilder
 import com.ingresse.sdk.base.IngresseCallback
 import com.ingresse.sdk.base.Response
 import com.ingresse.sdk.base.RetrofitCallback
@@ -20,18 +21,16 @@ import retrofit2.converter.gson.GsonConverterFactory
 import retrofit2.converter.scalars.ScalarsConverterFactory
 
 class UserService(private val client: IngresseClient) {
-    private var host = Host.API
     private var service: User
 
     private var mUserDataCall: Call<String>? = null
     private var mUpdateBasicInfosCall: Call<String>? = null
 
     init {
-        val adapter = Retrofit.Builder()
-                .addConverterFactory(ScalarsConverterFactory.create())
-                .addConverterFactory(GsonConverterFactory.create())
-                .baseUrl(URLBuilder(host, client.environment).build())
-                .build()
+        val adapter = RetrofitBuilder(
+            client = client,
+            hasGsonConverter = true)
+            .build()
 
         service = adapter.create(User::class.java)
     }
